Move mapper to util
