+++ .phpstorm.meta.php
@@ -2,4 +2,4 @@
 
 namespace PHPSTORM_META;
 
-override(\App\Services\Mapper::map(0, 1), type(1));
+override(\App\Services\Util\Mapper::map(0, 1), type(1));
