+++ src/Services/Util/Mapper.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Services;
+namespace App\Services\Util;
 
 class Mapper
 {
